feat: Add .gitattributes file to specify linguist-language
The recent code changes involve adding a new file `.gitattributes` to the repository. This file specifies that all `*.ipynb` files should be recognized as Python code by the Linguist language detection tool. This will ensure that the Jupyter Notebook files are correctly identified and highlighted as Python code in the repository.